requirements.txt
flask
tensorflow
gdown
pillow
numpy
